Adding L and R rays
Added L and R rays, along with Debug Rays for scene view, so that the Patron can also figure out if there is a bookcase on 30 degree angles each side. Also added two bools to determine, this way meaning that I can tell them to move towards the area where the space is clearer.